cli: Accept intermixed arguments and options
Currently the syntax of a command requires all the arguments to appear
before all the options. This complicates things when the value of an
option is required by the autocompletion mechanism in order to decide
which are the acceptable arguments. This patch changes the parsing rules
so that arguments and options can be freely intermixed.

Change-Id: I8c1f6982e7beddb17a4668f5719294a8aaae3c2e
Bug-Url: https://bugzilla.redhat.com/1047513
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>