Revert "cli: mismatch between restore snapshot syntax and auto completion #1027298"
This reverts commit 870b0c as it introduces a regression: moving the
options to the end of the command moves them past the redirections and
pipes, and this causes a syntax error. This is the bug that the change
intended to fix:

  https://bugzilla.redhat.com/1047513

A later patch introduces changes in the syntax of commands allowing
mixing arguments and options. That fixes the above bug as well as the
regression.

Change-Id: I7b85120e6479c4be3dd2357ef8731cd1388b8885
Bug-Url: https://bugzilla.redhat.com/1047513
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>