cli: Change action help text to match syntax
For auto completion to work the parent identifiers of the resource have
to be provided before the action, as the set of possible actions depends
on the actual parent. This patch changes the help text so that it
explains this.

Change-Id: Ib3d82efa4a7fbff4ff04cb9a2c0d7b9fec3670e8
Bug-Url: https://bugzilla.redhat.com/1047543
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 6528469db2435bb770a0a26edcdf4137a3435a72)