cli: Show HTTP error page only in debug mode
Currently when we try to connect to the engine and the connection fails
with an HTTP error code we print the error document to the output. This
is usually an HTML document, and not friendly for the user. This patch
changes the CLI so that it will only print the error document when the
-d option has been specified.

Change-Id: Ic6a3a1ec12f343038a6b0ddac2fe875b46841471
Bug-Url: https://bugzilla.redhat.com/869283
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>