cli: Don't require /api
Currently we perform an explicit validation of the URL of the API, and
this requires that the path component of the URL is exactly /api. This
doesn't work with the new URL schema of the engine, and won't work
either if the user decides use any other base URL. This patch removes
the validation of the path.

Change-Id: I71fac959498d17e1176ea16584e5b4d2c36ddbf4
Bug-Url: https://bugzilla.redhat.com/1069171
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 95314e74e4699b7c828f88c27d01ceaf1bc416ed)