cli: Don't write integer statistics as float
Currently when we write a an statistic value that is declared to be
integer we use the floating point representation used by default by
Python. This happens because all the statistics are represented as as
xs:decimal, regardless of their declared type, so the generateDS.py tool
converts them to float. This has the inconvenient that integer types are
displayed with a .0 suffix. For example:

  # show statistic memory.total --host-identifier myhost
  id                : 7816602b-c05c-3db7-a4da-3769f7ad8896
  name              : memory.total
  description       : Total memory
  host-id           : d3560912-f2ad-11e1-8f5f-001a4a169798
  type              : GAUGE
  unit              : BYTES
  values-type       : INTEGER
  values-value-datum: 33658241024.0

This patch changes the CLI so that it will convert statistics declared
as integer to Python long before rendering, so the .0 suffix will not be
present.

Change-Id: I21490ec50e3a7433331e981162de800c7d548f27
Bug-Url: https://bugzilla.redhat.com/855724
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>