cli: Update to generateDS.py 2.12a
Starting with version 3.5.0.0 of the Python SDK the version of
generateDS.py used to generate the code will be 2.12.a. One of the
changes in this version of generateDS.py is that it doesn't import the
datetime class like it used to do. In version 2.9a it imported it as
follows:

  from from datetime import datetime, tzinfo, timedelta

In version 2.12a it imports it in a different way:

  import datetime as datetime_

We currently depend on the name used by the generated code, so the CLI
doesn't work with the new version of generateDS.py. To fix that problem
and avoid it in the future this patch changes the CLI so that it doesn't
depend on how generateDS.py imports the class.

In addition this patch also updates the setup.py and the RPM spec to
require version 3.5.0.0 of the Python SDK.

Change-Id: Ia1b1529e9428a0b90f914b582f12153c809bdc5b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit a11d0807d87bb361500f351324e33a60349d7d8e)