cli: No regular expression for IP, UUID or number
Currently the command line lexer uses regular expressions to match IP
addresses, UUIDs and numbers. But in the parser grammar there is no
production that accepts those terminal symbols and doesn't accept also
plain words. The only special treatment is that numeric values are
converted to "int".

This behaviour means that some words that happen to start like IP
addresses, UUIDS or numbers are split into two tokens. For example, the
following command line:

  update host 10.10.10.10a --name 10.10.10.10

Will be split into the following tokens:

  WORD: update
  WORD: host
  IPADDR: 10.10.10.10
  WORD: a
  OPTION: --name
  IPADDR: 10.10.10.10

But it should have been as follows:

  WORD: update
  WORD: host
  WORD: 10.10.10.10a
  OPTION: --name
  IPADDR: 10.10.10.10

To avoid this issue, and simplify the lexer, this patch removes these
regular expressions for IP addresses, UUIDs and numbers, and replaces
them with a single expression that matches words. The action for this
expression will check if the value is actually an integer and convert it
to "int".

Change-Id: If670144751fdb9564380a12fb73596ecbaa47ee9
Bug-Url: https://bugzilla.redhat.com/1110366
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>