cli: Calculate singulars and plurals correctly
Currently when we need to calculate singulars and plurals we just remove
or add a final "s". This doesn't work for names like "SchedulingPolicy"
and "SchedulingPolicies". This patch improves the calculations so that
they will work correctly for this type of names.

Change-Id: I4c7849dadd1161bfd435c7ae79d196c08b3e7e69
Bug-Url: https://bugzilla.redhat.com/1144110
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit cbcf83df88901eca7de9b99fbad90f0b7c549301)