cli: Don't depend on optparse
Currently the dependency on the "optparse" module is deep in the
implementation of the CLI. This causes issues, like unexpected exit from
the "connect" command when passing incorrect options. To avoid these
issues this patch isolates the use of the "optparse" module to the main
program, so that the rest of the code will just receive the already
parsed options.

Change-Id: If3d71de72f929125229e8c8d0e86df5239fad07d
Bug-Url: https://bugzilla.redhat.com/1135550
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>