cli: Add support for --type-name options
Currently when a parent object needs to be specified in a command the
only alternative is to use the "--type-identifier" option. This option
will locate the parent object that matches the given identifier if it is
an UUID, or the name otherwise. The problem is that some parent objects
have identifiers that aren't UUIDs, so this won't be able to locate
them. To avoid this issue this patch introduces new "--type-name"
options that can be used in the same places that the "--type-identifier"
options are used. When using the existing "--type-identifier" option the
parent object will be located only by id, and when using the
"--type-name" option it will be located only by name, thus elimiating
any potential ambiguity.

Change-Id: I438e0d5e3c29302fda63e0f0ab01e87a6841e10e
Bug-Url: https://bugzilla.redhat.com/1177367
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>