cli: Try id, name and alias for show
The fix for bug 1148941 (commit 4b654a2) introduced a regression in the
way that the "show" command locates objects, so that objects can't be
located by name. For example, the following command will fail to locate
the a virtual machine, even if it exists:

  # show vm myvm

This patch fixes that regression, so that when given an object
identifier the CLI will try to locate it first by identifier, then by
name and finally by alias (for disks).

Change-Id: I8d197c042e83939470aea93405546a25bdf3f0c0
Related-To: https://bugzilla.redhat.com/1148941
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>