cli: Add "parent" prefix to parent identifier options
Currently the way to specify the identifier of a parent entity is to use
the "--*-identifier" or "--*-name" options. These option, specially the
"name" option, may conflict with other options used to populate the
request. For example, when adding a storage domain the "--host-name"
option is used to indicate what host name should perform the operation,
but the CLI assumes that it is used to specify the parent host, which
doesn't make sense. This patch changes the CLI so that it will require
the "parent" prefix in the option used to specify the parent name. The
option used to specify the parent identifier will also support the
"parent" prefix, but it will still work without it, for backwards
compatibility.

Change-Id: I4150389b6c87cc178865757f99a2e09e272c530f
Bug-Url: https://bugzilla.redhat.com/1213393
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>