cli: Try options from command line and settings
Currently the connect command doesn't try to get the options from the
command line, only from the settings. This is a regression introduced by
a previous patch that removed the dependency on the "optparse" module.
This patch fixes that issue.

Change-Id: I497a9269390fa62406099854cbebf273664c04a6
Bug-Url: https://bugzilla.redhat.com/1135544
Related: https://bugzilla.redhat.com/1135550
Related: https://gerrit.ovirt.org/36257
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>