cli: Support escape of double quote in strings
Currently strings support escape sequences, but they aren't replaced by
the corresponding characters. For example, in the following command:

  # update vm myvv --name "My \"VM\""

The name of the VM will be set to exactly this:

  My \"VM\"

Note that the backslash isn't removed.

This makes it impossible to use strings that represent data that
contain double quotes, in particular it makes it impossible to use XML
documents, which are needed when setting the OVF representation of a VM.

This patch changes the CLI so that the escape character preceding the
double quote will be removed, thus the above command will result in the
following name:

  My "VM"

Change-Id: Ied53ac577efa3395438892de535c35cd068a12bd
Bug-Url: https://bugzilla.redhat.com/1211670
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>