cli: Check if Spice port is available
Currently when the Spice port isn't available the CLI generates a
command line that includes "None" as the value for the "-p" option. This
causes a failure of the "spicec" command. To solve that problem this
patch changes the CLI so that it won't generate the "-p" option in that
case.

Change-Id: I36556b0dd5ee0cc794dbb7867719e4f90edaf924
Bug-Url: https://bugzilla.redhat.com/1234382
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>