cli: Generate parse tables during build
Currently the parse tables used by ply are generated during the RPM
installation, or, for non RPM platforms, during runtime. This causes
warnings because the user running the CLI may not have permissions to
write those parse tables. To avoid those warnings this patch changes the
CLI build process so that the parse tables are generated from the
"setup.py" script and included in the binary distribution.

Change-Id: I933a56b3ccbf5bc0e39feed9a2c61be89970dd84
Bug-Url: https://bugzilla.redhat.com/1255409
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>