cli: Don't check for old version of the SDK
This patch removes a check of the SDK version that the CLI used to do in
order to perform verification of credentials. This is no longer
necessary because the CLI explicitly requires (in both setup.py and in
the .spec file) a version of the SDK where this isn't needed. In
addition the way the version comparison is done triggers a warning
message in some versions of the Python setuptools package.

Change-Id: Ia83622fc122d30d9f0b2492755a0caff262c89ba
Related: https://bugzilla.redhat.com/1255409
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>