cli: Process correctly "--parent..." options
Currently when the CLI tries to find a parent collection it assumes that
the option used is "--type-identifier", but it can also be
"--parent-type-identifier" and "--parent-type-name". As a result when
the "--parent..." options are used autocompletion doesn't show the
correct options. To avoid this issue this patch changes the CLI so that
it uses the methods that check this in a more generic way.

Change-Id: I42fb362c59d4baad5968c1ac0df15bbe0b7dab8a
Bug-Url: https://bugzilla.redhat.com/1282832
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 1dcf01ef247ead15b6c49f7b6efc95f8ceac7d41)