cli: Generate remote viewer config using UTF-8
Currently the CLI generates the remote-viewer configuration files using
the default encoding used by Python, which happens to be "ascii". One of
the pieces of information that is written to these files is the subject
of the certificate of the host. This is retrieved from the engine, and
it may contain characters that aren't part of the ASCII character set.
When this happens Python tries to encode text using the "ascii" codec,
and that generates an exception. To avoid that this patch changes the
CLI so that these configuration files are generated using an Unicode
template and encoding the result explicitly using the "utf-8" codec
instead of the default.

Change-Id: Ib85d5134da87f7d901f4a7ff3bedfb7d7a3af628
Bug-Url: https://bugzilla.redhat.com/807696
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 78f509f2e27faf440d212f863b1c7081d3d6f82d)