cli: Improve the "no console found" messages
Currently the messages used to indicate that no console viewer was found
are the same for VNC and SPICE, and they are't very friendly as they
don't provide a hint about the binary that needs to be installed. This
patch improves the messages so that they include a hint about the binary
that needs to be installed.

When the graphics protocol is SPICE the error message will be the
following:

  a SPICE viewer was not found, please install a SPICE viewer (e.g.
  remote-viewer) first.

When the graphics protocol is VNC the error message will be the
following:

  a VNC viewer was not found, please install a VNC viewer (e.g.
  remote-viewer, or vncviewer) first.

Change-Id: I9952a2576a8544be712123e8ce0b6b92babacda0
Related-To: https://bugzilla.redhat.com/show_bug.cgi?id=807696#c13
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>