cli: Iterate all permutations of parent identifiers
When the CLI processes a command that contains multiple parent
identifiers it should process all the permutations in order to find the
correct collection. For example, for the following command:

  # add vm --id 123 \
  --parent-cluster-identifier 456 \
  --parent-affinitygroup-identifier 789

The CLI should try the following permutations:

  1) --parent-cluster-identifier, --parent-affinitygroup-identifier
  2) --parent-affinitygroup-identifier, --parent-cluster-identifier

Which should result in trying the following API URLs:

  1) /clusters/456/affinitygroups/789
  2) /affinitygroups/789/clusters/456

The first will work, the second will not, as it doesn't exist.

This is theory, but in practice the CLI doesn't try these permutations
correctly, only the first one.

In addition the ordering of the permutations depends on the names of the
options. For example, if the options don't have the "parent" prefix,
then the order of the permutations will be different, and the first one,
the only actually tried, will also be different.

This patch fixes the code that tries these permutations, so that it will
actually try all of them, till one succeeds.

Change-Id: I9ea62ef10f8cf04d9e7ee54b2ac420555bd567be
Bug-Url: https://bugzilla.redhat.com/1335844
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>