cli: Try to locate objects using id, name and alias
Currently when the CLI tries to locate an object it tries to do it by
id, by name and by alias, but only if the --id, --name or --alias
options are part of the command line. This is not correct for commands
like "show", which don't require any of those options, as the identifier
of the object is passed as an argument, not as an option. The result is
that the "show" command doesn't work. To fix that issue this patch
changes the CLI so that it will always try the id, name and alias,
regardless of the options. If one try fails, for whatever the reason, it
will be ignored and the next will be used.

Change-Id: I1ef5dbefe35cc9ac82884df7d51f912ca171e776
Bug-Url: https://bugzilla.redhat.com/1356607
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>