cli: Ignore "id" and "name" options for the "show" command
Currently auto-completion for the "show" command includes the "id" and
"name" options. This is wrong, as these options aren't used to specify
the object to be displayed. The reason they are included in
auto-completion is that the underlying Python SDK method is "get", and
that method usually has both the "id" and "name" parameters. We simply
and blindly extract the parameters from the method and assume that they
are valid options. This patch changes the CLI so that it ignores these
parameters, thus they won't appear in the auto-completion list.

Change-Id: I068686a6c600d0aac582adc239439c5f7dba0f13
Bug-Url: https://bugzilla.redhat.com/1358227
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>