Use alternate PKIResourceServlet to fetch CA certificate
Since oVirt 4.0 there is no ca.crt located in root of the apache,
but there is running servlet to fetch the CA certificate. This
patch changes the behaviour so we first try to fetch the CA
certificate from the root of the apache and if that don't exists,
we try to fetch it from new PKIResourceServlet servlet.

Change-Id: I5da25e4be36bee56c5ec9ee0dabead574dedada1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1362222
Signed-off-by: Ondra Machacek <omachace@redhat.com>