cli: Base should be None when there is no base
Currently when a base object can't be found the 'resolve_base' method returns
the starting point for the search, the 'connection' object. This is wrong, it
should return 'None', as that is what callers expect. This patch fixes that
issue.

Change-Id: I971624874935ef7342cd7fb193783893fc7cdc84
Bug-Url: https://bugzilla.redhat.com/1383735
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 42080df2d9405e3f9577106b588c8a93b2220008)